pyproject.toml
[tool.poetry]
name = "chat-search"
version = "0.1.3"
description = ""
authors = ["Hemslo Wang <hemslo.wang@gmail.com>"]
readme = "README.md"
packages = [
    { include = "app" },
]

[tool.poetry.dependencies]
beautifulsoup4 = "^4.12.3"
huggingface-hub = "^0.22.2"
langchain = "^0.1.16"
langchain-community = "^0.0.34"
langchain-openai = "^0.1.3"
langserve = { extras = ["server"], version = "^0.0.51" }
lark = "^1.1.9"
lxml = "^5.2.1"
opentelemetry-distro = { extras = ["oltp"], version = "^0.45b0" }
opentelemetry-exporter-otlp = "^1.24.0"
opentelemetry-instrumentation = "^0.45b0"
opentelemetry-instrumentation-aiohttp-client = "0.45b0"
opentelemetry-instrumentation-asgi = "0.45b0"
opentelemetry-instrumentation-aws-lambda = "^0.45b0"
opentelemetry-instrumentation-dbapi = "^0.45b0"
opentelemetry-instrumentation-fastapi = "0.45b0"
opentelemetry-instrumentation-grpc = "^0.45b0"
opentelemetry-instrumentation-httpx = "0.45b0"
opentelemetry-instrumentation-logging = "0.45b0"
opentelemetry-instrumentation-redis = "0.45b0"
opentelemetry-instrumentation-requests = "0.45b0"
opentelemetry-instrumentation-sqlalchemy = "0.45b0"
opentelemetry-instrumentation-sqlite3 = "0.45b0"
opentelemetry-instrumentation-tortoiseorm = "0.45b0"
opentelemetry-instrumentation-urllib = "0.45b0"
opentelemetry-instrumentation-urllib3 = "0.45b0"
opentelemetry-instrumentation-wsgi = "^0.45b0"
pydantic = "<2"
pyroscope-io = "^0.8.7"
python = "^3.11"
python-dotenv = "^1.0.0"
redis = "^5.0.4"
tqdm = "^4.66.4"
uvicorn = "^0.23.2"

[tool.poetry.group.dev.dependencies]
langchain-cli = "^0.0.21"
ruff = "^0.4.3"
redisvl = "^0.0.7"

[tool.poetry.scripts]
crawl = "app.client.crawl:main"

[tool.ruff.lint]
extend-select = ["I"]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
